+++ src/main/java/pl/project/bot/dto/TradeDTO.java
@@ -4,24 +4,25 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.ta4j.core.Trade;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
 @NoArgsConstructor
 public class TradeDTO {
-    private Double openPrice;
-    private Double closePrice;
+    private BigDecimal openPrice;
+    private BigDecimal closePrice;
     private Integer amount;
-    private Double profitLose;
+    private BigDecimal profitLose;
     private LocalDateTime dateOpen;
     private LocalDateTime dateClose;
-    private Double stopLoss;
-    private Double takeProfit;
+    private BigDecimal stopLoss;
+    private BigDecimal takeProfit;
     private String comment;
     private Trade.TradeType type;
 
-    public TradeDTO(Double openPrice, Integer amount, LocalDateTime dateOpen, Double stopLoss,
-                    Double takeProfit, Trade.TradeType type) {
+    public TradeDTO(BigDecimal openPrice, Integer amount, LocalDateTime dateOpen, BigDecimal stopLoss,
+                    BigDecimal takeProfit, Trade.TradeType type) {
         this.openPrice = openPrice;
         this.amount = amount;
         this.dateOpen = dateOpen;
